requirements.txt
numpy==1.21.5
Pillow==8.4.0
streamlit==1.3.0
tensorflow-cpu==2.7.0